lab 3
Understanding breakpoint and how to use them in debugging, in this lab we replaced a constant and instead put the data into memory then grabbed that from into a register. The issue was the code was using the memory address instead of the actual data it needed to be using. 
It's an easy one line fix just load the actual data from the memory address after retrieving the memory address
ldr r3, =DELAY_TIME
ldr r3, [r3]